+++ NewLife.Remoting/Models/CommandReplyModel.cs
@@ -1,12 +1,10 @@
 ﻿namespace NewLife.Remoting.Models;
 
-/// <summary>
-/// 命令响应模型
-/// </summary>
+/// <summary>命令响应模型</summary>
 public class CommandReplyModel
 {
     /// <summary>服务编号</summary>
-    public Int32 Id { get; set; }
+    public Int64 Id { get; set; }
 
     /// <summary>状态</summary>
     public CommandStatus Status { get; set; }
